backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
twilio==8.7.0
python-dotenv==1.0.0
gunicorn==21.2.0
langdetect==1.0.9
gtts==2.4.0 